moving classmethods to their own utils file
